back/requirements.txt
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
bitarray==0.8.1
certifi==2017.4.17
cffi==1.10.0
characteristic==14.3.0
chardet==3.0.4
click==6.7
colorama==0.3.9
constantly==15.1.0
cryptography==1.9
cssselect==1.0.1
Flask==0.12.2
Flask-SQLAlchemy==2.2
html5lib==0.999999999
hyperlink==17.2.1
idna==2.5
incremental==17.5.0
itsdangerous==0.24
Jinja2==2.9.6
lxml==3.8.0
MarkupSafe==1.0
olefile==0.44
parsel==1.2.0
Pillow==4.2.0
ply==3.10
pyasn1==0.2.3
pyasn1-modules==0.0.9
pybloom-live==2.2.0
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.1.0
queuelib==1.4.2
requests==2.18.1
Scrapy==1.4.0
service-identity==17.0.0
sh==1.12.14
six==1.10.0
SQLAlchemy==1.1.11
Twisted==17.5.0
typing==3.6.1
update==0.4.4
urllib3==1.21.1
w3lib==1.17.0
webencodings==0.5.1
Werkzeug==0.12.2
zope.interface==4.4.2
